requirements.txt
requests
tavily-python
openai
langchain>=0.3.0
langchain-ollama
langchain-core
langchain-community
langchain-openai
langchain-chroma
langgraph>=0.2.0
chromadb
streamlit
python-dotenv
python-docx
marker-pdf
jq
